+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 * Add abstract class Aihrus_Settings
 * Add abstract class Aihrus_Widget
 * Add is_true, terms, url validate cases
+* Add method `create_link( $link )`
+* Add method `create_nonce( $action )`
+* Add method `verify_nonce( $nonce, $action )`
 * Add shortcode id helpers
 * Allow reset_defaults in widget
 * Rename no_code to show_code